+++ src/opcodes/Div.cpp
@@ -32,8 +32,8 @@ Instruction* Div::CreateInstruction(Memory::MemoryOffset& memLoc, Processor* pro
 
 	Instruction* newDiv = 0;
 
-	if((*opLoc == DIV_MOD8 || *opLoc == DIV_MOD16) && 
-			(((*(opLoc + 1) & 0x38) >> 3) == DIV_SUB_OPCODE)) {
+	if((*opLoc == DIV_MOD8 || *opLoc == DIV_MOD16) &&
+			((unsigned int)((*(opLoc + 1) & 0x38) >> 3) == DIV_SUB_OPCODE)) {
 		unsigned int size = *opLoc == DIV_MOD8 ? 1 : 2;
 		Operand* dst = ModrmOperand::GetModrmOperand(proc, opLoc, ModrmOperand::MOD, size);
 		snprintf(buf, 65, "DIV %s", dst->GetDisasm().c_str());
